refactor(librarian): move install and version cmds
